requirements.txt
numpy
pandas
matplotlib
torch
torchvision
Pillow
jupyter
